Keyboardz and Mousez class
Decided it was better to have a the fiddler object contain separate keyboard and mouse objects within it.

The idea is that reading from the keyboard or mouse will output the chord and or mouse direction. The fiddler object will switch between different operational modes depending on a defined input.